agents/test_agent/requirements.txt
nats-py>=2.0.0
pytest>=7.0.0
pytest-json-report>=1.4.0
python-dotenv>=1.0.0